flask/requirements.txt
certifi==2021.10.8
chardet==3.0.4
click==8.0.4
Flask==2.0.3
idna==2.10
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dotenv==0.15.0
requests==2.23.0
urllib3==1.25.11
Werkzeug==2.0.3
